+++ src/energy/tests/compare_with_excel.rs
@@ -3,7 +3,7 @@ use crate::sectors::SectorsResult;
 use crate::sectors::SectorsRawValues;
 use crate::result::Results;
 
-fn assert(a: [f32; 4], b: [f32; 4]){
+fn assert(a: [f64; 4], b: [f64; 4]){
     assert_relative_eq!(a[0], b[0], max_relative=0.3);
     assert_relative_eq!(a[1], b[1], max_relative=0.3);
     assert_relative_eq!(a[2], b[2], max_relative=0.3);
